Introducing the Register class
