+++ devices.py
@@ -3,6 +3,7 @@ from abc import ABC, abstractmethod
 from scipy.spatial.distance import pdist
 
 from channels import Raman, Rydberg
+from register import Register
 
 
 class PasqalDevice(ABC):
@@ -12,16 +13,21 @@ class PasqalDevice(ABC):
     following this template.
 
     Args:
-        qubits (dict): Dictionary with the qubit names as keys and their
-            position coordinates as values (e.g. {'q0':(2, -1, 0), ...}).
+        qubits (dict, Register): A dictionary or a Register class instance with
+            all the qubits' names and respective positions in the array.
     """
 
     def __init__(self, qubits):
-        if not isinstance(qubits, dict):
-            raise TypeError("The qubits have to be stored in a dictionary.")
+        if isinstance(qubits, dict):
+            register = Register(qubits)
+        elif isinstance(qubits, Register):
+            register = qubits
+        else:
+            raise TypeError("The qubits must be a in a dict or Register class "
+                            "instance.")
 
-        self.check_array(list(qubits.values()))
-        self._qubits = qubits
+        self.check_array(list(register.qubits.values()))
+        self._register = register
 
     @property
     @abstractmethod
@@ -67,7 +73,7 @@ class PasqalDevice(ABC):
     @property
     def qubits(self):
         """The dictionary of qubit names and their positions."""
-        return dict(self._qubits)
+        return self._register.qubits
 
     def check_array(self, atoms):
         if len(atoms) > self.max_atom_num:
@@ -88,76 +94,6 @@ class PasqalDevice(ABC):
             raise ValueError("All qubits must be at most {}um away from the "
                              "center of the array.".format(
                                                     self.max_radial_distance))
-
-    @classmethod
-    def rectangle(cls, rows, columns, spacing=4, prefix=None):
-        """Initializes the device with the qubits in a rectangular array.
-
-        Args:
-            rows (int): Number of rows.
-            columns (int): Number of columns.
-
-        Keyword args:
-            spacing(float): The distance between neighbouring qubits in um.
-            prefix (str): The prefix for the qubit ids. If defined, each qubit
-                id starts with the prefix, followed by an int from 0 to N-1
-                (e.g. prefix='q' -> IDs: 'q0', 'q1', 'q2', ...)
-        """
-        coords = np.array([(x, y) for x in range(columns)
-                           for y in range(rows)], dtype=float) * spacing
-
-        return cls(cls._coords_to_qubits(coords, prefix=prefix))
-
-    @classmethod
-    def square(cls, side, spacing=4, prefix=None):
-        """Initializes the device with the qubits in a square array.
-
-        Args:
-            side (int): Side of the square in number of qubits.
-
-        Keyword args:
-            spacing(float): The distance between neighbouring qubits in um.
-            prefix (str): The prefix for the qubit ids. If defined, each qubit
-                id starts with the prefix, followed by an int from 0 to N-1
-                (e.g. prefix='q' -> IDs: 'q0', 'q1', 'q2', ...).
-        """
-        return cls.rectangle(side, side, spacing=spacing, prefix=prefix)
-
-    @classmethod
-    def triangular_lattice(cls, rows, atoms_per_row, spacing=4, prefix=None):
-        """Initializes the device with the qubits in a triangular lattice.
-
-        Initializes the qubits in a triangular lattice pattern, more
-        specifically a triangular lattice with horizontal rows, meaning the
-        triangles are pointing up and down.
-
-        Args:
-            rows (int): Number of rows.
-            atoms_per_row (int): Number of atoms per row.
-
-        Keyword args:
-            spacing(float): The distance between neighbouring qubits in um.
-            prefix (str): The prefix for the qubit ids. If defined, each qubit
-                id starts with the prefix, followed by an int from 0 to N-1
-                (e.g. prefix='q' -> IDs: 'q0', 'q1', 'q2', ...).
-        """
-        coords = np.array([(x, y) for x in range(atoms_per_row)
-                           for y in range(rows)], dtype=float)
-        coords[:, 0] += 0.5 * np.mod(coords[:, 1], 2)
-        coords[:, 1] *= np.sqrt(3) / 2
-        coords *= spacing
-
-        return cls(cls._coords_to_qubits(coords, prefix=prefix))
-
-    @staticmethod
-    def _coords_to_qubits(coords, prefix=None):
-        """Centers a coords array and returns the respective qubits dict."""
-        coords -= np.mean(coords, axis=0)      # Centers the array
-        if prefix is not None:
-            pre = str(prefix)
-            return {pre+str(i): pos for i, pos in enumerate(coords)}
-        else:
-            return dict(enumerate(coords))
 
 
 class Chadoq2(PasqalDevice):
